Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 2938-separate-black-and-white-balls/2938-separate-black-and-white-balls.cpp
@@ -6,7 +6,7 @@ public:
     typedef vector<vl> vll;
 
     long long minimumSteps(string s) {
-        ll answer=0;
+        int answer=0;
         int n=s.size();
         int goal=n-1;
         for(int i=n-1;0<=i;i--){
